Fixes #45 - NAPALM Endpoint Support
Adds support for the napalm detail endpoint by adding a `.napalm`
property to the Device object that returns a `RODetailEndpoint`
object. Since it's a RODetailEndpoint `create()` raises a
`NotImplementedError`.

`list()` is the only use-case and now accepts kwargs. Namely, `method`
to pass the proper url parameter to the napalm endpoint. Docs with
examples are included.