Fixes #141 - ChangeObject str return
Creates a custom model for ChangeObjects so their __str__ can be set to
a field that exists on that model, request_id. Though not very human
friendly at the momement, I hope we can use `changed_object_type` to
load the proper model for objects in `changed_object` and we can use
their __str__ methods instead.