Issue #212 Accessing related resources via parent attribute uses correct models
Fix an issue whereby retrieving an object directly, e.g:
```
interface = client.dcim.interfaces.get(1)
```
and then accessing a related object from that object:
```
device = interface.device
```
would result in device being of a generic Record() type and the Endpoint
being that of the original interface object.  This is dangerous because:
```
device.delete()
```
would actually call delete against the interface and not the device.

This change, parses the url value of related objects to infer the App()
and record model that is appropriate to generate a new record of the
correct type with its proper endpoint.